+++ app/src/main/java/com/knowhow/android/picturewithai/MainActivity.java
@@ -21,6 +21,7 @@ import android.os.Bundle;
 import android.provider.MediaStore;
 import android.util.Log;
 import android.view.View;
+import android.view.WindowManager;
 import android.widget.Button;
 import android.widget.ImageView;
 import android.widget.ProgressBar;
@@ -57,23 +58,17 @@ import com.knowhow.android.picturewithai.utils.FileUtil;
 
 public class MainActivity extends AppCompatActivity {
 
-    final int REQUEST_EXTERNAL_STORAGE = 100;
 
-    HerokuService service;
-    ServiceInterface serviceInterface;
-    List<Uri> files = new ArrayList<>();
-    Dialog dialog;
+
+
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
+        getWindow().setFlags(WindowManager.LayoutParams.FLAG_FULLSCREEN,WindowManager.LayoutParams.FLAG_FULLSCREEN);
         setContentView(R.layout.activity_main);
 
-        Retrofit retrofit = new Retrofit.Builder()
-                .baseUrl("https://photo-with-ai.herokuapp.com/")
-                .build();
 
-        service = retrofit.create(HerokuService.class);
 
 
         //uploadImage();
@@ -115,239 +110,5 @@ public class MainActivity extends AppCompatActivity {
 
 
 
-    //===== Upload files to server
-    public void uploadImages(){
 
-
-
-        List<MultipartBody.Part> list = new ArrayList<>();
-
-        for (Uri uri:files) {
-
-            Log.i("uris",uri.getPath());
-
-            list.add(prepareFilePart("file", uri));
-        }
-
-        serviceInterface = ApiConstants.getClient().create(ServiceInterface.class);
-
-
-        Call<ResponseBody> call = serviceInterface.uploadNewsFeedImages(list);
-
-        ProgressBar progress = (ProgressBar) findViewById(R.id.progress) ;
-        progress.setVisibility(View.VISIBLE);
-
-
-        call.enqueue(new Callback<ResponseBody>() {
-
-
-
-            @Override
-            public void onResponse(Call<ResponseBody> call, Response<ResponseBody> response) {
-
-                System.out.println("hello");
-                System.out.println(response.code());
-                try {
-                    Log.i("hh", String.valueOf(response));
-
-                    ResponseBody addMediaModel = response.body();
-
-                    System.out.println(response.body().string());
-                    progress.setVisibility(View.GONE);
-
-                    //Intent intent = new Intent(MainActivity.this, BestPicture.class);
-                    //intent.putExtra("data", "Test Popup");
-                    //startActivity(intent);
-                    //dialog=new Dialog(MainActivity.this);
-                    //dialog.setContentView(R.layout.activity_best_picture);
-
-                    //View bestpicture=findViewById(R.id.best);
-
-
-
-                    File f = new File(String.valueOf(files.get(0)));
-                    Drawable d = Drawable.createFromPath(f.getAbsolutePath());
-                    
-                    Log.d("ii2", f.getAbsolutePath());
-
-                    dialog.show();
-
-
-                }
-                catch (Exception e){
-                    Log.d("Exception","|=>"+e.getMessage());
-//
-                }
-            }
-
-            @Override
-            public void onFailure(Call<ResponseBody> call, Throwable t) {
-
-                Log.i("my",t.getMessage());
-            }
-        });
-    }
-
-    private String getRealPathFromURI(Uri contentURI) {
-        Cursor cursor = getContentResolver().query(contentURI, null, null, null, null);
-        if (cursor == null) { // Source is Dropbox or other similar local file path
-            return contentURI.getPath();
-        } else {
-            cursor.moveToFirst();
-            int idx = cursor.getColumnIndex(MediaStore.Images.ImageColumns.DATA);
-            return cursor.getString(idx);
-        }
-    }
-
-    @NonNull
-    private MultipartBody.Part prepareFilePart(String partName, Uri fileUri) {
-
-        File file = new File(fileUri.getPath());
-        Log.i("here is error",file.getAbsolutePath());
-        // create RequestBody instance from file
-
-        RequestBody requestFile =
-                RequestBody.create(
-                        MediaType.parse("image/*"),
-                        file);
-
-        // MultipartBody.Part is used to send also the actual file name
-        return MultipartBody.Part.createFormData(partName, file.getName(), requestFile);
-
-
-    }
-
-    public void testHeroku(){
-        Call<ResponseBody> call = service.hello();
-        call.enqueue(new Callback<ResponseBody>() {
-            @Override
-            public void onResponse(Call<ResponseBody> _,
-                                   Response<ResponseBody> response) {
-                System.out.println("hello2");
-                System.out.println(response.code());
-            }
-
-            @Override
-            public void onFailure(Call<ResponseBody> _, Throwable t) {
-                System.out.println("good");
-            }
-        });
-    }
-
-    public void launchGalleryIntent() {
-        Intent intent = new Intent(Intent.ACTION_GET_CONTENT);
-        intent.putExtra(Intent.EXTRA_ALLOW_MULTIPLE, true);
-        intent.setType("image/*");
-        startActivityForResult(intent, REQUEST_EXTERNAL_STORAGE);
-    }
-
-
-
-
-
-    @Override
-    public void onRequestPermissionsResult(int requestCode,
-                                           String[] permissions, int[] grantResults) {
-        switch (requestCode) {
-            case REQUEST_EXTERNAL_STORAGE: {
-                // If request is cancelled, the result arrays are empty.
-                if (grantResults.length > 0
-                        && grantResults[0] == PackageManager.PERMISSION_GRANTED) {
-                    // permission was granted, yay! Do the
-                    // contacts-related task you need to do.
-                    launchGalleryIntent();
-                } else {
-                    // permission denied, boo! Disable the
-                    // functionality that depends on this permission.
-                }
-                return;
-            }
-
-            // other 'case' lines to check for other
-            // permissions this app might request.
-        }
-    }
-
-
-    @Override
-    protected void onActivityResult(int requestCode, int resultCode, @Nullable Intent data) {
-        super.onActivityResult(requestCode, resultCode, data);
-        if (requestCode == REQUEST_EXTERNAL_STORAGE && resultCode == RESULT_OK) {
-
-            //final ImageView imageView = findViewById(R.id.imageview);
-            final List<Bitmap> bitmaps = new ArrayList<>();
-            ClipData clipData = data.getClipData();
-
-            if (clipData != null) {
-                //multiple images selecetd
-                for (int i = 0; i < clipData.getItemCount(); i++) {
-                    Uri img = clipData.getItemAt(i).getUri();
-
-
-
-
-                    String imgPath = FileUtil.getPath(MainActivity.this,img);
-
-                    files.add(Uri.parse(imgPath));
-
-
-                    Log.d("URI", img.toString());
-                    try {
-                        InputStream inputStream = getContentResolver().openInputStream(img);
-                        Bitmap bitmap = BitmapFactory.decodeStream(inputStream);
-                        bitmaps.add(bitmap);
-                    } catch (FileNotFoundException e) {
-                        e.printStackTrace();
-                    }
-                }
-
-            } else {
-                //single image selected
-
-                Uri img = data.getData();
-
-
-                String imgPath = FileUtil.getPath(MainActivity.this,img);
-
-                files.add(Uri.parse(imgPath));
-
-                //uploadImages();
-                try {
-                    InputStream inputStream = getContentResolver().openInputStream(img);
-                    Bitmap bitmap = BitmapFactory.decodeStream(inputStream);
-                    bitmaps.add(bitmap);
-                } catch (FileNotFoundException e) {
-                    e.printStackTrace();
-                }
-
-            }
-            uploadImages();
-
-//            new Thread(new Runnable() {
-//                @Override
-//                public void run() {
-//                    for (final Bitmap b : bitmaps) {
-//                        runOnUiThread(new Runnable() {
-//                            @Override
-//                            public void run() {
-//                                imageView.setImageBitmap(b);
-//
-//                                //uploadImage();
-//                                //fileUpload(new File(filePath));
-//                                //uploadImage();
-//                                //testHeroku();
-//
-//                            }
-//                        });
-//
-//                        try {
-//                            Thread.sleep(3000);
-//                        } catch (InterruptedException e) {
-//                            e.printStackTrace();
-//                        }
-//                    }
-//                }
-//            }).start();
-        }
-    }
 }
