+++ app/src/main/res/layout/activity_best_picture.xml
@@ -1,60 +1,76 @@
 <?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:background="#eeeeee"
-    android:orientation="vertical"
-    android:layout_width="300dp"
-    android:layout_height="wrap_content">
+<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
 
-    <!-- 타이틀바 -->
-    <LinearLayout
-        android:orientation="horizontal"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        >
+    android:background="#F6F6F6">
 
 
-
-        <ImageView
-            android:id="@+id/best"
-            android:layout_width="200dp"
-            android:layout_height="200dp"
-            android:layout_alignParentLeft="true"
-            android:layout_marginLeft="43dp"
-            android:layout_alignParentBottom="true"
-            android:layout_gravity="center_horizontal"
-            android:layout_marginBottom="50dp"
-            />
-
-    </LinearLayout>
-    <!-- //end 타이틀바 -->
-
-
-
-    <!-- Notice -->
 
     <View
-        android:background="#66bdbdbd"
-        android:layout_width="match_parent"
-        android:layout_height="1dp" />
+        android:id="@+id/photo_with_"
+        android:background="@drawable/mainmenutitle"
+        android:layout_width="200dp"
+        android:layout_height="60dp"
 
-    <!-- 닫기 버튼 -->
-    <LinearLayout
-        android:orientation="horizontal"
-        android:gravity="center"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content">
-        <Button
-            android:text="확인"
-            android:textSize="15sp"
-            android:textColor="#ff7a00"
-            android:padding="16dp"
-            android:gravity="center"
-            android:background="#00000000"
-            android:layout_width="match_parent"
-            android:layout_height="53dp"
-            android:onClick="mOnClose"/>
-    </LinearLayout>
-    <!--// 닫기 버튼 -->
-</LinearLayout>
+        android:layout_centerInParent="true"
+        android:layout_alignParentTop="true"
+        android:layout_marginTop="30dp"
+        android:lineSpacingExtra="2sp"
+        android:gravity="center_vertical"/>
+
+
+    <ImageView
+        android:id="@+id/bestPicture"
+        android:layout_width="360dp"
+        android:layout_height="443dp"
+        android:layout_centerInParent="true"
+        android:layout_alignParentTop="true"
+        android:layout_marginTop="80dp"
+
+        />
+
+
+
+    <ImageButton
+        android:layout_width="50dp"
+        android:layout_height="50dp"
+        android:layout_centerInParent="true"
+        android:layout_alignParentTop="true"
+        android:layout_marginTop="600dp"
+        android:src="@drawable/share"
+        android:id="@+id/btn"/>
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+    <ProgressBar
+        android:id="@+id/progress"
+        android:layout_width="100dp"
+        android:layout_height="100dp"
+        android:layout_centerInParent="true"
+        android:visibility="gone"
+        />
+
+
+
+
+
+</RelativeLayout>
